fix: corrige tipo da variável id

+++ src/main/java/com/mirasystems/payroll/exception/EmployeeNotFoundException.java
@@ -4,7 +4,7 @@ public class EmployeeNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = -4331656123393763680L;
 
-	public EmployeeNotFoundException(Long id) {
+	public EmployeeNotFoundException(Integer id) {
 		super("Could not find employee " + id);
 	}
 }
